Move merge-related files into their own module
