+++ test/unit/data/test_data_merge_image.py
@@ -3,7 +3,7 @@ from typing import Tuple, Iterable
 import numpy as np
 
 from fedot.core.data.data import OutputData
-from fedot.core.data.data_merger import DataMerger
+from fedot.core.data.merge.data_merger import DataMerger
 from fedot.core.repository.dataset_types import DataTypesEnum
 from fedot.core.repository.tasks import Task, TaskTypesEnum
 from fedot.core.data.array_utilities import atleast_4d
