+++ test/unit/data/test_data_merge_ts.py
@@ -1,12 +1,10 @@
 from itertools import product
-from typing import Union
 
 import numpy as np
 import pytest
 
-from fedot.core.data.data import InputData, OutputData
-from fedot.core.data.data_merger import DataMerger
-from fedot.core.data.supplementary_data import SupplementaryData
+from fedot.core.data.data import OutputData
+from fedot.core.data.merge.data_merger import DataMerger
 from fedot.core.pipelines.pipeline_builder import PipelineBuilder
 from fedot.core.repository.dataset_types import DataTypesEnum
 from fedot.core.repository.tasks import Task, TaskTypesEnum, TsForecastingParams
